can-logos: add logo for linux-can GitHub organization
GitHub allows to view the pages in 'dark mode'. The resistors of the
CAN harness are ending directly at the logo border which looks pretty
ugly with a black background in the dark mode.

The added canvas with rounded corners has a 2.5% grey filling which
makes the background invisible in the 'light mode' and centers the
logo with some margin (for the resistors) in the 'dark mode'

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>